Make the parser use crate-level definitions
