+++ hir/Cargo.toml
@@ -22,6 +22,7 @@ miden-assembly.workspace = true
 miden-diagnostics.workspace = true
 miden-hir-symbol.workspace = true
 miden-hir-type.workspace = true
+miden-parsing = "0.1"
 petgraph = "0.6"
 paste.workspace = true
 rustc-hash.workspace = true
