+++ hir/src/parser/ast/block.rs
@@ -1,3 +1,4 @@
+use crate::{Ident, Type, };
 use super::*;
 
 const INDENT: &str = "    ";
