+++ hir/src/parser/ast/functions.rs
@@ -1,3 +1,4 @@
+use crate::{ArgumentPurpose, ArgumentExtension, CallConv, FunctionIdent, Type, };
 use super::*;
 
 /// The possible visibilities of a function
